+++ x64barebones/Userland/SampleCodeModule/test_processes.c
@@ -32,7 +32,7 @@ int64_t test_processes(int64_t argc, char *argv[]) {
 
     // Create max_processes processes
     for (rq = 0; rq < max_processes; rq++) {
-      p_rqs[rq].pid = sys_register_child_process((uint64_t)&endless_loop, 1, 1, argvAux);
+      p_rqs[rq].pid = sys_register_child_process((uint64_t)&endless_loop, 1, 1, (uint64_t) argvAux);
       if (p_rqs[rq].pid == -1) {
         printf("test_processes: ERROR creating process\n");
         return -1;
